Класс V8Object вынесен в отдельные файлы
Обновлены проекты code::blocks и cmake
Добавлен файл с общими константами классов
