+++ src/ctool1cd/CMakeLists.txt
@@ -5,10 +5,11 @@ set (tool_VERSION_MAJOR 1)
 set (tool_VERSION_MINOR 0)
 
 set (TOOL_SOURCES cTool_1CD_Main.cpp ParseCommandLine.cpp MessageRegistration.cpp Class_1CD.cpp
-	Common.cpp ConfigStorage.cpp Parse_tree.cpp APIcfBase.cpp TempStream.cpp Base64.cpp UZLib.cpp Messenger.cpp)
+	Common.cpp ConfigStorage.cpp Parse_tree.cpp APIcfBase.cpp TempStream.cpp Base64.cpp UZLib.cpp Messenger.cpp
+	V8Object.cpp)
 set (TOOL_HEADERS cTool_1CD_Main.h ParseCommandLine.h MessageRegistration.h Class_1CD.h
 	Common.h ConfigStorage.h Parse_tree.h APIcfBase.h TempStream.h Base64.h UZLib.h Messenger.h ErrorCode.h
-	db_ver.h NodeTypes.h)
+	db_ver.h NodeTypes.h V8Object.h Constants.h)
 
 if (CMAKE_VERSION VERSION_LESS "3.1")
 	if(CMAKE_COMPILER_IS_GNUCXX)
